platform/x86: add support for devices with Silead touchscreens
On ACPI based tablets, the ACPI touchscreen node only contains info on
the gpio and the irq, and is missing any info on the axis. This info is
expected to be built into the tablet model specific version of the driver
shipped with the os-image for the device.

Add support for getting the missing info from a table built into the
driver, using dmi data to identify which entry of the table to use and
add info for the CUBE iwork8 Air and Jumper EZpad mini3 tablets on which
this code was tested / developed.

BugLink: https://bugzilla.kernel.org/show_bug.cgi?id=187531
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
[dmitry.torokhov@gmail.com: Move to platform/x86]
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>
Acked-by: Andy Shevchenko <andy.shevchenko@gmail.com>
---
Changes in v2:
-Put the dmi code in a separate silead_dmi.c file
-Use device_add_properties to add the info
Changes in v3 (Dmitry Torokhov):
-Move the silead_dmi code to drivers/platform/x86
Changes in v4:
-Minor spelling and code-style fixes based on review from
 Andy Shevchenko
Changes in v5:
-Use has_acpi_companion()
-Drop .ident from dmi_id structs, it is not used